Check the TOC is up-to-date for PRs
A Github action will run on pushes, regenerating the TOC and complaining
if there are differences with the version from the checked-out tree.
The generator no longer uses `os.walk` because dirctory traversal order
is not guaranteed, so we implement the traversal ourselves to guarantee
the TOC will be generated with elements in a predictable order.